chore: refactor egui editor tabs into their own files
