+++ crates/bevy_animation_graph_editor/src/ui/editor_windows/mod.rs
@@ -1,2 +1,6 @@
 pub mod debugger;
+pub mod event_sender;
+pub mod graph_editor;
+pub mod preview_hierarchy;
 pub mod scene_preview;
+pub mod scene_preview_errors;
